LayoutTests/js/array-species-different-globalobjects-expected.txt

PASS result instanceof Array is true
PASS result instanceof Array is true
PASS result instanceof Array is true
PASS result instanceof Array is true
PASS result instanceof Array is true
PASS result instanceof Array is true
PASS successfullyParsed is true

TEST COMPLETE
